requirements.txt
# Daily Beltway — Python dependencies
# Install with: pip install -r requirements.txt

# CLI framework
click>=8.1,<9.0

# YAML config parsing
pyyaml>=6.0,<7.0

# .env file loading
python-dotenv>=1.0,<2.0

# Slug generation (URL-safe strings)
python-slugify>=8.0,<9.0

# Dropbox API upload (optional — only needed if DROPBOX_ACCESS_TOKEN is set)
dropbox>=11.36,<13.0

# Testing
pytest>=7.0,<9.0
